feat: divide the main page into components
