+++ packages/nextjs/aizktec/src/components/footer/Footer.tsx
@@ -1,0 +1,11 @@
+const Footer = () => {
+    return (
+        <footer className="bg-black text-gray-500 flex justify-center p-4 mt-10 max-w-4xl">
+            <div className="p-4 bg-customGray rounded-lg border border-white/30 text-center w-full">
+                AI-ZK-TEC: Empowering users with confidence in AI through advanced privacy and compliance.
+            </div>
+        </footer>
+    );
+};
+
+export default Footer;
